Make builds reproducible.
Preserve any timestamps and support the `${project.build.outputTimestamp}` which
sets the timestamps for all entries created by the plugin in the jar.

Note that the timestamp may be different from entries that have been lifted from
an inlined jar. This is due to MJAR-300 and will stay this way until the bug
is fixed (the timestamps generated by the plugin are correct, the ones in the
jar generated by the maven-jar-plugin are not).